[PBL6-25] - Add service to manager Job and TypeOfJob

See merge request bk-pbl/pbl6!20

+++ be/src/PBL6.CasualManager.Domain/JobInfos/JobInfo.cs
@@ -19,7 +19,7 @@ namespace PBL6.CasualManager.JobInfos
         public string Description { get; set; }
 
         [Required]
-        public float Prices { get; set; }
+        public int Prices { get; set; }
 
         public Guid TypeOfJobId { get; set; }
